CI: Fix K8sEventDetectionTest skip for OpenShift

+++ qa-tests-backend/src/test/groovy/K8sEventDetectionTest.groovy
@@ -9,6 +9,7 @@ import services.PolicyService
 import util.Env
 
 import org.junit.Assume
+import spock.lang.IgnoreIf
 import spock.lang.Retry
 import spock.lang.Tag
 import spock.lang.Unroll
@@ -34,6 +35,10 @@ class K8sEventDetectionTest extends BaseSpecification {
     static final private String CLONED_KUBECTL_EXEC_POLICY_NAME = "CLONED: Kubernetes Actions: Exec into Pod"
 
     def setupSpec() {
+        if (Env.mustGetOrchestratorType() == OrchestratorTypes.OPENSHIFT) {
+            // K8s event detection is not supported on OpenShift.
+            return
+        }
         orchestrator.batchCreateDeployments(DEPLOYMENTS)
         for (Deployment deployment : DEPLOYMENTS) {
             assert Services.waitForDeployment(deployment)
@@ -55,6 +60,10 @@ class K8sEventDetectionTest extends BaseSpecification {
     }
 
     def cleanupSpec() {
+        if (Env.mustGetOrchestratorType() == OrchestratorTypes.OPENSHIFT) {
+            // K8s event detection is not supported on OpenShift.
+            return
+        }
         for (def deployment: DEPLOYMENTS) {
             orchestrator.deleteDeployment(deployment)
         }
@@ -123,12 +132,11 @@ class K8sEventDetectionTest extends BaseSpecification {
     @Tag("BAT")
     @Tag("RUNTIME")
     @Tag("K8sEvents")
+    // K8s event detection is currently not supported on OpenShift.
+    @IgnoreIf({ Env.mustGetOrchestratorType() == OrchestratorTypes.OPENSHIFT })
     def "Verify k8s exec detection into #execIntoDeploymentNames with addl groups #additionalPolicyGroups"() {
         when:
         "Create the deployments, modify the policy, exec into them"
-        // K8s event detection is currently not supported on OpenShift.
-        Assume.assumeTrue(Env.mustGetOrchestratorType() != OrchestratorTypes.OPENSHIFT)
-
         def originalPolicy = Services.getPolicyByName(CLONED_KUBECTL_EXEC_POLICY_NAME)
         assert originalPolicy != null && originalPolicy.getName() == CLONED_KUBECTL_EXEC_POLICY_NAME
